+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrated Lennard Jones example to PyTorch Geometric.
 - Migrated physicsnemo.utils.sdf.signed_distance_field to a static return,
   torch-only interface.  It also now works on distributed meshes and input fields.
+- Refactored DiTBlock to be more modular
+- Added NATTEN 2D neighborhood attention backend for DiTBlock
 
 ### Deprecated